requirements.txt
chitchat_dataset
halo
numpy>=1.18
tensorflow-gpu>=1.15.0,<2.0.0
t5==0.3.0
